Now soft deletes if the entity should be soft deleted

+++ module/Olcs/Db/src/Service/ServiceAbstract.php
@@ -194,7 +194,12 @@ abstract class ServiceAbstract implements OlcsRestServerInterface
             return false;
         }
 
-        $this->getEntityManager()->remove($entity);
+        if ($this->canSoftDelete()) {
+            $entity->setIsDeleted(true);
+            $this->dbPersist($entity);
+        } else {
+            $this->getEntityManager()->remove($entity);
+        }
         $this->dbFlush();
 
         return true;
